zephyr: cros_system: npcx: fix chip reset cause checking
In CrOS EC, the system driver uses SCRATCH and VCC1_RST_STS bits in
RSTCTL register to check the reset cause. With these two bits, the
driver can identify the system's reset comes from the hardware
(ex: Power-ON or VCC1_RST pin) or non-hardware (ex: sysjump to RW).
In Zephyr, the system driver only uses the VCC1_RST_STS bit to check
the reset cause. As a result, the driver treats a non-hardware reset
as the hardware reset, which might cause a problem like bug:226094431.

BUG=b:226094431
BRANCH=none
TEST=pass "zmake testall --clobber"
TEST=the system succesfully sysjumps to RW after it wakeups from the
hibernation.

Signed-off-by: Jun Lin <CHLin56@nuvoton.com>
Change-Id: I15fd8f3b3c064e718716a05c26de602d59b38aa3