zephyr: shim: decouple syswork support data from task_ctx_data
This is an alternative implementation of:

ff45017216 Zephyr: Create sysworkq entry in task shim

The original one was adding a full new task context structure for the
sysworkq task, but that was not making any use of the "struct k_thread"
field, which takes quite a bit of RAM.

This version creates a new statically allocated structure just for the
sysworkq task, and only allocates the required fields, at the expense of
handling the TASK_ID_SYSWORKQ case exxplicitly in various places.

Net resource tradeoff on brya is 224 bytes of RAM for 104 bytes of
flash.

BRANCH=none
BUG=none
TEST=build and run on brya
TEST=cq dry run

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: If2b3a53875900c26f2357a08ff3f9beb7c63c5e7