flash_fp_mcu: add a delay before doing power cycling
We recently found that the fpmcu couldn't be probed from the spi bus
after doing the power cycle in the brya platform. After analysis, the
reason should be related to the RC delay in the PWR_EN which makes the
power cycle hit some critical things that need the power sustain.
Based on this finding, we add a delay to avoid the problem.

BUG=b:191110371, b:189788793, b:186695335
BRANCH=None
TEST=`test_that -b brya firmware_Fingerprint.RDP1` passed

Change-Id: If0dd3bbd7c1cb3c928aa1c82c5171fd9e90f5f1f
Signed-off-by: Zhuohao Lee <zhuohao@chromium.org>