board/hyperdebug: Add board for OpenTitan development
HyperDebug is a USB-connected device, which is to be used in
conjunction with an OpenTitan development board during development of
OpenTitan firmware.  HyperDebug will serve a purpose similar to
uServo, that is forward several UART ports via USB, and allow GPIO and
SPI/I2C manipulation via USB from the development host computer.

This CL adds the first iteration of board/hyperdebug, with declaration
for UART, SPI and I2C forwarding, and names for each GPIO pin based on
its physical location on on of two 70-pin DIL expansion ports on the
HyperDebug board.

BUG=b:192262089
TEST=User HyperDebug in place of UltraDebug for uploading firmware \
	and sending TPM commands to AndreiBoard
BRANCH=none

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I44b26d2890d0d4ed44d6bcd59a4de992a594c03f