servo_v4p1: add command to send data over USART4 interface
To control the keyboard emulator on the servos ATMEGA32U4 chip, the
user needs to be able to send data over the STM32F072's USART4
interface. This CL adds a new command to allow sending data to the
keyboard emulator using the board's USART4 interface.

BUG=b:226209216
TEST=Used the new command to send HID key press and release events from
the servo to a connected DUT.
BRANCH=None

Signed-off-by: Jameson Thies <jthies@google.com>
Change-Id: Ib24f0228ef49ed86b222c21632489c9143c15f71