zephyr: move USB mux configuration into the device tree
Create a separated usb_muxes.c file in the shim layer to define the
usb_muxes array. Additional USB muxes in a chain are created in the same
file.

Common properties of the usb_mux DTS node are in cros-ec,usb-mux.yaml
file.

USB muxes configuration of the Krabby is moved to the device tree.

BUG=b:227757117
TEST=zmake testall
TEST=check if usb_muxes structure is correct using gdb:
     gdb ./build/zephyr/krabby/output/zephyr.ro.elf
     p usb_muxes
BRANCH=none

Signed-off-by: Tomasz Michalec <tm@semihalf.com>
Change-Id: I6cb26404b5cb44c6fce29f4176ebb6707af71d21