RGBKBD: Handle EC_CMD_PWM_SET_DUTY
In a regular design, a keyboard backlight is connected to a PWM pin of
an EC. With an RGB keyboard, the brightness control is part of the RGB
keyboard functionality. Thus, EC_CMD_PWM_SET_DUTY with
EC_PWM_TYPE_KB_LIGHT should by handled by an RGB keyboard API to control
the keyboard backlight brightness.

BUG=b:228525798
BRANCH=None
TEST=Taniks

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: Id8f23ba67728f03664673834be0903e0f593e744