TCPMV2: Update sending Alerts and add a Status response
Currently, USB PD alert messages are built based on whether a port is a
source or a sink. If the port is a source, the EC assumes an OCP event
and if the port is a sink the EC assumes an OVP event. To support other
types of alert messages, the EC must change how they are built. This CL
adds alert data object (ADO) information to the PE state and functions
in usb_common to set the ADO and send alerts on one or all ports. It
will allow the alert message to be defined where it is triggered.

Additionally, partners may respond to an Alert message with a Get_Status
message. This CL adds support for responding to Get_Status messages with
a Status message.

BUG=b:226228461, b:227236917
TEST=triggered alert message from power state change and observed
Alert->Get_Status->Status transactions with OWC TBT4 dock on a twinkie.
BRANCH=None

Signed-off-by: Jameson Thies <jthies@google.com>

Change-Id: Ib4ad0ead9eaaff792c17836bcb0061789dcccaf7