common: Make IS_ENABLED work at low optimization
This allows for the use of "-Og", which reduces compiler optimization
to a debugger friendly level.
https://gcc.gnu.org/onlinedocs/gcc/Optimize-Options.html

The previous mechanism relies on compiler optimization
to not check the IS_ENABLED_BAD_ARGS() call when the
"if (!__undefined)" branch will never happen, thus not emitting
the error associated with the function.

When you turn off optimization, all references (and branches)
are checked, thus the error is always emitted.

This new method uses _Static_assert to raise a build time error message.

BRANCH=none
BUG=b:172212718
TEST=make BOARD=$BOARD -j V=1 CFLAGS_DEBUG_CHIP="-ggdb -Og" \
         build/$BOARD/ec.bin
TEST=compare_builds.sh -b all
     # Use objdump and meld to ensure asm differences
     # are functionally equivalent.

Change-Id: I8884b80e05ead10e78f0c8be05158590618bf3a8
Signed-off-by: Craig Hesling <hesling@chromium.org>