corsola: enable DRP until XHCI inited.
Previously we issued a hard reset signal to cycle the VBUS after
XHCI inited. The hard reset was happened when AP at coreboot stage,
but somehow the outdated hard reset event was carried into kernel
causing a wrong type-c state.

To prevent that, we force the port at SINK state until the XHCI
inited, and then enable DRP toggling so we don't need to cycle the
VBUS and issue hard reset.

BUG=b:224471170
TEST=Plug external monitor and boot device, the MUX state is correct
     at kernel.
BRANCH=none

Change-Id: I848e8e086a50b1eea8b70befdfb24599346c5743
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>