zmake: Cleanup all the pylint warnings in zmake
Added type hints in several places either to correct a warning or to
cause one. For example, if pylint can't tell the type of an inherited
method, it will report unused params, but if the params match exactly
the superclass, it will not. Some cases were because I tried to find all
references to a class or function, and the IDE couldn't find them.

Moved loose logging functions in multiproc into class methods of
LogWriter.

Refactored pack_firmware and _get_max_image_bytes to avoid problems with
the various packers taking different numbers of params and subclasses
not implemented abstract methods.

Stop disabling no-self-use and fix them instead.

Remove JobClient.run() because it's never called.
Remove JobServer, because it does nothing and is not used as a marker.

Fixed lots of small warnings also.

Disabled most of the too-many-* warnings in specific functions.

Fixed the imports in .pylintrc, as it was finding the emerged zmake, not
the local one.

Disabled wrong-import-order since isort and pylint don't agree.

BRANCH=None
BUG=b:217969201
TEST=zmake -j8 test -a
zephyr/zmake/run_tests.sh
find zephyr/zmake -name '*.py' -print0 | xargs -0 cros lint

Signed-off-by: Jeremy Bettis <jbettis@google.com>
Change-Id: I4f1b3a9fa5881fe7bcc4a6065d326dd36701b5ef