zephyr: tasks: add structured support for extra tasks
Code that runs in Zephyr defined threads need some extra structure
support, this has been done with an ad-hoc structure for the sysworkq
thread, but it's needed again for running stuff in the main thread.

Adding some structure and support code to support this use case with
minimal code duplication.

BRANCH=none
BUG=b:227772790
TEST=build and run on brya

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: I180a5c9cf56083cf6c9ebe23794446b171ee597c