zephyr: add yaml file for charger chips
Add the binding yaml files for all charger chips used in Zephyr. It will
be needed for generating the chg_chips array based on DTS, not defining
in board-specific code.
The yaml files for ISL923x and ISL9241 chips were already present, so
just adjust them to the desired usage.

BUG=b:228237412
TEST=zmake testall
BRANCH=main

Signed-off-by: Dawid Niedzwiecki <dn@semihalf.com>
Change-Id: I0e0120d28f03c5d1ba5fb1bd9bbb5a1e878af44f