zephyr: tasks: run HOSTCMD in main thread
Add an option to run the host command loop in the main thread, taking
over the entire thread and reusing its thread support code and stack.

This is still optional and automatically disabled on tests as test code
actually needs main() to run the test code itself.

The code ensures that the main thread is switched to the correct
priority for the HOSTCMD code and changes the thread name as well.

BRANCH=none
BUG=b:223044986
TEST=zmake testall
TEST=build and run on brya

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: Ib2b3ba3d8a6fb883876ecdd432068ca3300e5344