zephyr: Kconfig: fix few choice symbol format
Choice entries expects a symbol (or nothing) as the argument. The
symbol is only used to have choices in multiple files (which we don't)
and for documentation referencing.

This renames few instances of descriptions where there should be a
symbol around various Kconfig files, and cleans up various
choice/endchoice definition cosmetic inconsistencies.

BUG=none
BRANCH=none
TEST=ninja menuconfig

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Link: https://www.kernel.org/doc/html/latest/kbuild/kconfig-language.html#kconfig-syntax
Change-Id: I040bc9008d6c97bf96d8cf4f37384504ccda78b1