corsola: do not reset aux_path immediately
We combined the DP_HPD pin and AUX_PATH configuration together.
However, when DP_HPD disabled, it could be just a HPD debounce
for generating HPD IRQ, and in this case we shouldn't reset the
AUX path immediately. We fix this by deferring the aux path reset
for 1ms and re-check.

BUG=b:228289594
TEST=plug a DP dongle at C1 and ensure it can output.
BRANCH=main

Change-Id: I8094da14a89382a36a691790a46b2758414f7628
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>