baseboard/mtscp-rv32i: add non-cacheable dram region
This patch aims to achieve 3 goals,
1. align DRAM region with kernel driver
2. reserve space for panic data
3. add non-cacheable region

The reason of goal 1 is because the SCP kernel driver allocates a
5 MB region for SCP firmware image on the kernel reserved memory.
It's more clear if this 5 MB size limit information can be found on
SCP firmware.

The SCP kernel driver also allocates buffer on reserved
memory after this 5 MB space. This may cause data corruption if
SCP firmware uses the space overlapped with kernel buffers.
There's currently no issue happened because the 5 MB is big enough for
current usage. However, I think it's better to obviously address the
range for future.

Goal 2 and 3 are reverving space from the 5 MB region. The cacheable
dram usage on mt8195 is ~317KB. There's still a plenty of spaces for
panic data and non-cacheable region in this region.

BRANCH=none
BUG=b:193814857
TEST=make BOARD=cherry_scp
TEST=make BOARD=asurada_scp

Signed-off-by: Tinghan Shen <tinghan.shen@mediatek.com>
Change-Id: I5c76f748bc702c63d274141ab198aa23266de9a3