zephyr: set sysworkq priority explicitly from the task list
Now that we have an explicit task priority list, add an entry for the
sysworkq and use it to set the priority. This makes that task priority
setting somewhat more explicit relatively to the other EC tasks, ensure
that it's set correctly and make the build assert work properly.

Also drop one of the two build assert, which was redundant.

Threads looks like this on Brya after this:

22-04-19 11:38:51.576  0x200c5588 CHG_RAMP
22-04-19 11:38:51.576   options: 0x0, priority: 22 timeout: 1
...
22-04-19 11:38:51.589  0x200c7b20 sysworkq
22-04-19 11:38:51.589   options: 0x0, priority: 23 timeout: 311
...
22-04-19 11:38:51.603  0x200c7950 idle
22-04-19 11:38:51.603   options: 0x1, priority: 24 timeout: 0

BRANCH=none
BUG=none
TEST=kernel threads ec command on brya

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: I667cddbba9906e96bd9ab1ba8edab5103cc6c4b1