agah: add CONFIG_CHIPSET_X86_RSMRST_AFTER_S5
When we use C2D2 to program the AP firmware, we need to enable the
EN_S5_RAILS. However, the power sequence ic will deassert the
SEQ_EC_RSMRST_ODL and the EC will bypass the RSMRST to the PCH and
make the AP starting to run. To avoid the spi bus contention with the
C2D2 during ap spi flash, we add CONFIG_CHIPSET_X86_RSMRST_AFTER_S5
to avoid the RSMRST be passed to the PCH before the S5.

BUG=b:226438219
BRANCH=None
TEST=1. use C2D2 to flash the ap firmware
     2. use CCD to flash the pa firmware

Change-Id: Id9fd4ded4c90f22eb5f2b7d740dbdbfd0843d2d9
Signed-off-by: Zhuohao Lee <zhuohao@chromium.org>