nipperkin: Set DC prochot current thresh at init
Nipperkin devices are spiking current frequently exceeding the 4 A
threshold for over current prochot assertion on the ISL.  Testing
shows that setting the current threshold limit to ~6 A will minimize
the impact to the user.  Care should be taken not to exceed this value
on Nipperkin devices as the battery will cutoff if sustained current is
>= 6A for a duration of 10 mS.  Keeping the current threshold below 6A
gives the device a chance to assert prochot and avoid a DC cutoff.

BUG=b:216489130
TEST=Boot, verify that the ISL dc prochot threshold register is set to
5.888 A
BRANCH=guybrush

Signed-off-by: Jon Murphy <jpmurphy@google.com>
Change-Id: I7ed388ab535dad228c40fbe5bcb108679d985418