common/keyboard_8042: Add a method to pause the to_host queue
On devices that use eSPI VW for IRQs there currently exists a race
condition when waking using the keyboard. It takes time for the eSPI VW
channel to get initialized on boot. If the keyboard was pressed to wake
the system, the scan codes will get written into the 8042 output buffer.
The NPCX devices are configured to send an IRQ message when the state of
the output buffer changes. If the VW channel isn't setup then the
message is dropped. This results in the VW state getting out of sync
between the host and EC.

This CL introduces a method to pause the to_host queue and prevent the
scan codes from being written into the 8042 output buffer. Once the VW
channel is configured the queue can be resumed and the scan code can be
written to the output buffer. This closes the race condition and keeps
our VW state in sync.

By default I left the queue enabled because I wanted this CL to be a
no-op for platforms that don't have this problem.

BUG=b:218874489, b:160595155
BRANCH=guybrush, zork
TEST=Use zork and guybrush and verify eSPI VW are synchronized on S0i3
and S3 resume.

Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Change-Id: I703377c47f972c9a6d49323eb5f14bcbc7af572b