TCMPV2: Add state to send Status extended messages
When a USB PD partner sends a Get_Status message, the EC should respond
with a Status message. This CL adds the Status message response and
creates a new function in usb_pd_dpm which will be used to build the
Status Data Block (SDB).

BUG=b:227236917
TEST=triggered Status message and saw that it correctly sent data from a
modified dpm_get_status_msg function on a twinkie.
BRANCH=None

Signed-off-by: Jameson Thies <jthies@google.com>

Change-Id: Ifafa6e9ee7d9a62ef75e5c9e96606ffba64c98b0