common/keyboard_8042: Add 8042 kbw command
This command will allow sending scan codes to the host as if the
keyboard was sending them. kbw stands for keyboard write. In the future
we can add an auxw to write to the AUX port. This command is useful for
testing keyboard input without actually having to press the keys.

e.g., Send a tab down and tab up

    > 8042 kbw 0xf 0x8f

BUG=b:218874489
BRANCH=guybrush, zork
TEST=Send tab key using kbw command.

Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Change-Id: I3f93fcb0586a4bcc8c3ef26c23ce5dde877bb73d