zephyr: shim: set an placeholder period cell for pwms property
Update all pwms entries to set a period property. This is part of an
ongoing standardization effort in the PWM subsystem in Zephyr, for the
moment just set a "0" placeholder (it's not used by any high level
driver), will followup setting it to the traget value once we'll have
the frequency dt macro available, and then use them in the higher level
driver and deprecate the old frequency label properties on a later
stage.

BRANCH=none
BUG=b:230093078
TEST=zmake testall

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Cq-Depend: chromium:3605350, chromium:3605359
Change-Id: I38cc7ab4cbbb8a7e256a913e3c592aa7ffd55000