smart battery: Fix mfgacc access function reading wrong number of bytes
`sb_read_mfgacc` currently uses `sb_read_string`. This function reads a
null-terminated string. However, the data that is being read is not a
null-terminated string. This is causing N-1 bytes to be read instead of
N bytes. This commit adds an `sb_read_sized_block` function that is used
instead. Only PARAM_SAFETY_STATUS and PARAM_OPERATION_STATUS are read
with `sb_read_mfgacc` and neither of these values are a string.

BUG=b:228589840
TEST=Ran i2ctrace to verify that right number of bytes are now being
read.
BRANCH=guybrush

Change-Id: Iaedb4d28274bda8b6d1ee6d2021ebd04f83767f8
Signed-off-by: Robert Zieba <robertzieba@google.com>