it8xxx2: don't set system reset flags directly
The get_reset_cause method of the cros_system driver is meant to only
return the cause. Setting the system reset flags directly there (rather
than allowing the caller to adjust them) is incorrect and causes
confusing behavior like reporting multiple mutually-incompatible reset
causes in `sysinfo` command output.

The tests for causes are reordered to return the earliest event that
could have occurred, preferring power-on then reset pin then watchdog.
It's not generally possible to determine which reset actually occurred
because the registers appear to only be cleared when written by
software, so this assumes that a set bit indicates a reset of that form
occurred and hasn't yet been noted by the firmware.

The twiddling of the watchdog bit is removed from the function
completely, since it didn't do anything: system_set_reset_flags will
only set bits in the flags, so attempting to clear
EC_RESET_FLAG_WATCHDOG had no effect.

BUG=b:227972919
TEST=Reset causes on Nereid now correctly reflect the actual last reset
     in sysinfo command output and do not indicate multiple causes.
BRANCH=none

Change-Id: I28fd125dfb476ba451c07622478345dd3b61f3cd
Signed-off-by: Peter Marheine <pmarheine@chromium.org>