throttle_ap: Add option to gate PROCHOT based on C10
For some x86 boards, the PROCHOT signal is not valid once the AP enters
the C10 state.  Add an option to gate PROCHOT detection if the C10 state
is asserted. When the AP exits C10, the EC rechecks the PROCHOT state.

Note that only the Volteer baseboard enables the PROCHOT interrupt, so
it is the only board that is updated to call
throttle_ap_config_prochot().

All gpio.inc files that connect to throttle_ap_prochot_input_interrupt()
all use the pin name EC_PROCHOT_IN_L. Confirmed with these searches:
    $ grep -r "GPIO_INT.*throttle_ap_prochot_input_interrupt" . \
    | wc -l
    31
    $ grep -r "EC_PROCHOT_IN_L.*throttle_ap_prochot_input_interrupt" . \
    | wc -l
    31

    $ grep -r "gpio_enable_interrupt.*EC_PROCHOT_IN_L" .
    ./baseboard/volteer/power.c: \
        gpio_enable_interrupt(GPIO_EC_PROCHOT_IN_L);

BUG=b:185810479
BRANCH=volteer
TEST=make buildall

Signed-off-by: Keith Short <keithshort@chromium.org>
Change-Id: I73fb328675d9faade13fe0192570dc838de028a6