nissa: don't control HDMI sub-board S5 rails
The EN_S5_RAILS input to the HDMI sub-board is not actually connected to
anything (it is enabled as long as the S5 3.3V rail from the mainboard is
provided to the sub-board), so it doesn't need to be controlled by the EC
and doing so affects hibernate functionality because that pin is connected
to the Z-sleep circuit on the mainboard since it doubles as the USB-C
interrupt for the other sub-board options (allowing the EC to wake from
hibernation when a charger is connected).

If the EC is asserting EN_S5_RAILS to the sub-board when attempting to
hibernate, hibernate fails because it is latched back off by the asserted
signal (anytime the AP is in S5 or higher power state). Since hibernate
should work in S5 as well as G3, simply stop controlling that EC output.

BUG=b:226325993
TEST=Nereid sb_1 (SUB_USB_C1_INT_ODL) remains high when AP is on with
     sub-board 3 in use, and the sub-board remains powered (the USB port
     on it works).
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: Idd6c0a37249d1e85a8fd0ffc763a31cee452f9cc