nissa: enable PROCHOT monitoring
The EC_PROCHOT_ODL signal was connected to the
throttle_ap_prochot_input_interrupt, but the interrupt was never
enabled.

Configure the PROCHOT input and enable the interrupt, which will log to
the EC console when PROCHOT is detected. If the board provides fans,
then the fans are also enabled during PROCHOT conditions.

BUG=b:230783041
BRANCH=none
TEST=zmake testall

Signed-off-by: Keith Short <keithshort@chromium.org>
Change-Id: I686a684e86d3c9897e7c980e41c3d4a2995e0ccc