chargesplash: Add ectool command and display code
Add a new "chargesplash" subcommand to ectool, which enables printing
and modifying the chargesplash state, as well as show the display loop
using frecon.

The display is intentionally nothing but a text-based UI at the
moment.  Future changes will add pretty graphics and such, but the
core logic is here with this commit.

Right now, the display loop just uses polling at 250 ms to read the
charging state and power button state.  For a better user experience,
we should grab the power button input device and terminate early in
the future.  This could even enable us to slow the polling further, as
charging state is not something that changes at high frequency anyway.

BUG=b:228370390
BRANCH=none
TEST=with whole CL stack, use ghost image and activate chargesplash on
     brya

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I5135c81f34d1f53b12670287e27e34161c9cff6b