usbc: fix the VID and PID field in Battery_Capabilities
When responding to an invalid Get_Battery_Cap request,
Set VID:PID field to FFFF:0000 according to PD 3.1 spec below.

Also rearranged the logic to handle invalid battery ref in
a single place.

--

6.5.5.1

If the Battery Cap Ref field in the Get_Battery_Cap Message is Invalid,
this VID field Shall be 0xFFFF.

6.5.5.2

The following rules apply to the PID field. When the VID:
  <skipped...>
  - Is 0xFFFF the PID field Shall be set to 0x0000.

BUG=b:229700747
TEST=pass TEST.PD.PROT.PORT3.4
BRANCH=none

Signed-off-by: Ting Shen <phoenixshen@google.com>
Change-Id: Ie7e9fbfeba8d9e9980c03a5cef0773492e39733a