zephyr: Disable optimization in drivers test cov
If coverage is enabled, disable optimization. This doesn't really matter
in the chroot, but with gcc it causes much better results. The default
optimization seems to remove most return statements from the coverage,
but does not remove them in the herobrine build, so those lines appear
as uncovered.

BRANCH=None
BUG=b:190519131
TEST=zmake test -c in docker image & chroot

Signed-off-by: Jeremy Bettis <jbettis@google.com>
Change-Id: I0b316f0033b5a24c8820861c63fb4599da14c436