ap_pwrseq: Use the power button task to start the CPU
Use the power button task processing to leave the CPU
off or to start it, rather than automatically starting it.

Since CONFIG_BRINGUP is currently set, the default
state at the start is for the CPU to be off, so it
needs a power button press to start.

BUG=b:223113743
TEST=zmake build nivviks; flash & run
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: Ia2e1f4f4160222ab4ee9e1ad1740ec8dd1ca3ec0