ap_pwrseq: Use the power button task to start the CPU
Use the power button task processing to leave the CPU
off or to start it, rather than automatically starting it.

Run the check to discover the CPU state earlier so that
the power button init code can check the state of the CPU.

Refactor ap_power_force_shutdown() to avoid sending
the notification.

BUG=b:223113743
TEST=zmake build nivviks; flash & run
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: Id66a95390d054d8530e39dd6ce7088621eda6b86