nissa: Turn off CONFIG_BRINGUP
CONFIG_BRINGUP by default will not start the CPU, though
the current power sequence code starts automatically.
Turn off CONFIG_BRINGUP in preparation for a fix that
changes the way the power sequence code handles the
starting of the CPU (the power button logic should be used
to control the starting of the CPU).

BUG=b:223113743
TEST=zmake build nivviks; flash & run
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: I496bacf9f5e9da09260fd813aade84eca80dd57b