ap_ro_integrity_check: update hash until the bid type is saved
On some devices the board id flags are set to lock in the phase and the
board id type isn't set until the board is finalized. RO may be changed
until the board id type is written. Change the check from
board_id_is_erased to board_id_type_is_blank, so the factory can update
the AP RO hash until the board is finalized.

This is the same check we do in sn_bits. Try to read the board id and
then check the type. In the future, we may want to consolidate.

BUG=b:230430292
TEST=manual
	Clear the board id
	Set the hash
	python ap_ro_hash.py GBB
	gsctool -aA prints the digest

	Set the BID flags
	gsctool -ai 0xffffffff:0x1234

	Clear the hash
	gsctool -aH
	gsctool -aA
	get hash rc: 10 AP RO hash unprogrammed

	Set the hash
	python ap_ro_hash.py GBB
	gsctool -aA prints the digest

	Clear the hash
	gsctool -aH
	gsctool -aA
	get hash rc: 10 AP RO hash unprogrammed

	Set the BID type
	gsctool -ai $(cros_config / brand-code):0x1234

	Verify cr50 rejects setting the hash
	python ap_ro_hash.py GBB
	ERROR: Cr50 returned 7 (BID programmed)

Change-Id: I440ee84b3c86e16f027a8b9dcd51ea3031171ea1
Signed-off-by: Mary Ruthven <mruthven@chromium.org>